+++ packages/headless/src/index.ts
@@ -58,6 +58,7 @@ export * from './features/index';
 export * from './features/analytics/index';
 
 // Types & Helpers
+export {API_DATE_FORMAT} from './api/search/date/date-format';
 export {TestUtils, HighlightUtils};
 export type {Result} from './api/search/search/result';
 export type {FieldDescription} from './api/search/fields/fields-response';